@@ -1,55 +1,49 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
 
-namespace SimpleApi.Controllers
+namespace SimpleApi.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class WeatherForecastController : ControllerBase
 {
-    [ApiController]
-    [Route("[controller]")]
-    public class WeatherForecastController : ControllerBase
+    private static readonly string[] Summaries = new[]
     {
-        private static readonly string[] Summaries = new[]
-        {
-            "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
-        };
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    };
 
-        private readonly ILogger<WeatherForecastController> _logger;
+    private readonly ILogger<WeatherForecastController> _logger;
 
-        public WeatherForecastController(ILogger<WeatherForecastController> logger)
+    public WeatherForecastController(ILogger<WeatherForecastController> logger)
+    {
+        _logger = logger;
+    }
+
+    [HttpGet(Name = "GetWeatherForecast")]
+    public IEnumerable<WeatherForecast> Get()
+    {
+        var rng = new Random();
+        
+        int temperatureC = rng.Next(-20, 55);
+        var summaryId = rng.Next(Summaries.Length);
+
+        switch (summaryId)
         {
-            _logger = logger;
+            case 0: case 9:
+            _logger.LogError("WeatherForecast: extreme weather");
+            break;
+            case 1: case 2: case 7: case 8:
+            _logger.LogWarning("WeatherForecast: severe weather");
+            break;
+            default:
+            _logger.LogInformation("WeatherForecast: mild weather");
+            break;
         }
-
-        [HttpGet]
-        public IEnumerable<WeatherForecast> Get()
+        return Enumerable.Range(1, 1).Select(index => new WeatherForecast
         {
-            var rng = new Random();
-            
-            int temperatureC = rng.Next(-20, 55);
-            var summaryId = rng.Next(Summaries.Length);
-
-            switch (summaryId)
-            {
-                case 0: case 9:
-                _logger.LogError("WeatherForecast: extreme weather");
-                break;
-                case 1: case 2: case 7: case 8:
-                _logger.LogWarning("WeatherForecast: severe weather");
-                break;
-                default:
-                _logger.LogInformation("WeatherForecast: mild weather");
-                break;
-            }
-            return Enumerable.Range(1, 1).Select(index => new WeatherForecast
-            {
-                Date = DateTime.Now.AddDays(index),
-                TemperatureC = temperatureC,
-                Summary = Summaries[summaryId]
-            })
-            .ToArray();
-        }
+            Date = DateTime.Now.AddDays(index),
+            TemperatureC = temperatureC,
+            Summary = Summaries[summaryId]
+        })
+        .ToArray();
     }
 }
