@@ -1,11 +1,9 @@
 ﻿@using System.Security.Principal
 
 <ul class="navbar-nav">
-@if (User.Identity.IsAuthenticated)
+@if (User.Identity?.IsAuthenticated == true)
 {
-        <li class="nav-item">
-            <span class="navbar-text text-dark">Hello @User.Identity.Name!</span>
-        </li>
+        <span class="navbar-text text-dark">Hello @User.Identity?.Name!</span>
         <li class="nav-item">
             <a class="nav-link text-dark" asp-area="MicrosoftIdentity" asp-controller="Account" asp-action="SignOut">Sign out</a>
         </li>
